move internals around
